Part of work with free controller

+++ src/main/java/our/slonbot/model/Player.java
@@ -1,17 +1,15 @@
 package our.slonbot.model;
 
-import our.slonbot.connection.IDataWorker;
-
 public class Player {
     public long id;
     public AppType appType;
-    public int appId;
+    public long appId;
     public String name;
     public long exp;
     public int level;
     public int money;
 
-    public Player(long id, AppType appType, int appId, String name, long exp, int level, int money) {
+    public Player(long id, AppType appType, long appId, String name, long exp, int level, int money) {
         this.id = id;
         this.appType = appType;
         this.appId = appId;
